implement guest order creation with entity updates

+++ src/main/java/com/abed/perfumeshop/coupon/dto/CouponResponse.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -21,7 +22,7 @@ public class CouponResponse {
 
     private String discountType;
 
-    private Double discountValue;
+    private BigDecimal discountValue;
 
     private LocalDateTime expiryDate;
